renamed setdit by set_ditgrid
